backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
google-generativeai>=0.8
openai>=1.0
langchain-text-splitters>=0.2
sentence-transformers>=2.2
PyPDF2>=3.0
python-docx>=1.0
python-dotenv>=1.0
supabase>=2.0
psycopg2-binary>=2.9
whitenoise>=6.5
gunicorn>=21.2
dj-database-url>=2.1
pgvector>=0.3
PyJWT>=2.8
cryptography>=41.0
